Dedinition of the models to be used in the application and addition of pictures in the assets folder for testing/development
A folder /lib/models has been created to contain the models of the application. For this commit the models created are:
- Order
- Order details
- Payment
- Payment type
- Payment validity
- Product
- Product category
- Shippng info
- Shipping type
- User
Additionally product tiles' widget responsible for displaying the products have been instantiated and a function getProductTiles have been created to generate the products tiles from the product list. for the moment they just display empty containers.
In this commit there is also a static list of products in /lib/models/product.dart file which can be used in testing/development while waiting for linking to the backend.