This commit provides addidional navigation routes, and a functional drawer
A new manner is Introduced to manage sessions where by upon routes generation a SellamAppSession widget is returned with the child parameter which is the desired route/page.
The drawer Items :
 - Home
 - Favorites
 - Notification
are now all operational as well as their routes/pages.
A filter dropdown has been added to the HomePage() to enable filtering of products.