Implementation of the Orders page and order details page.
Addition of a static variable, "route" to all pages to ease navigation and upadate of the app.dart file.
Modifications and update of some models and initialisation of list of of the corresponding model type.
Definition of a signIn() and signOut() methods in the SellamAppSession widget.
Implementation of Navigation routes for the oders and order details pages.